+++ gems/smithy-cbor/CHANGELOG.md
@@ -1,0 +1,2 @@
+Unreleased Changes
+------------------
