+++ gems/smithy-cbor/VERSION
@@ -1,0 +1,1 @@
+1.0.0.pre0
