+++ gems/smithy-cbor/sig/smithy-cbor/codec.rbs
@@ -1,0 +1,8 @@
+module Smithy
+  module CBOR
+    module Codec
+      def serialize: (Schema::Shapes::Shape shape, untyped data) -> (nil | String)
+      def deserialize: (Schema::Shapes::Shape shape, String bytes, ?Struct[untyped] target) -> (Hash[untyped, untyped] | ::Struct[untyped] | Schema::Union)
+    end
+  end
+end
