+++ gems/smithy-cbor/spec/decode-error-tests.json
