+++ gems/smithy-cbor/spec/decode-success-tests.json
