+++ gems/smithy-client/lib/smithy-client.rb
@@ -33,11 +33,6 @@ require_relative 'smithy-client/http/request'
 require_relative 'smithy-client/net_http/connection_pool'
 require_relative 'smithy-client/net_http/handler'
 
-# codecs
-
-require_relative 'smithy-client/cbor'
-require_relative 'smithy-client/codecs'
-
 # stubbing
 
 require_relative 'smithy-client/stubs'
