--- gems/smithy-client/lib/smithy-client/codecs.rb
@@ -1,11 +1,0 @@
-# frozen_string_literal: true
-
-require_relative 'codecs/cbor'
-
-module Smithy
-  module Client
-    # Contains different serialization and deserialization formats
-    module Codecs
-    end
-  end
-end
