--- gems/smithy-client/sig/smithy-client/codec/cbor.rbs
@@ -1,13 +1,0 @@
-module Smithy
-  module Client
-    module Codecs
-      class CBOR
-        def initialize: (?::Hash[untyped, untyped] options) -> void
-
-        def serialize: (Schema::Shapes::Shape shape, untyped data) -> (nil | String)
-
-        def deserialize: (Schema::Shapes::Shape shape, String bytes, ?Struct[untyped] target) -> (Hash[untyped, untyped] | ::Struct[untyped] | Schema::Union)
-      end
-    end
-  end
-end
