+++ gems/smithy-schema/lib/smithy-schema.rb
@@ -5,8 +5,8 @@ require_relative 'smithy-schema/structure'
 require_relative 'smithy-schema/union'
 
 module Smithy
-  # Base module for Smithy model classes.
-  module Model
+  # Base module for Smithy schema classes.
+  module Schema
     VERSION = File.read(File.expand_path('../VERSION', __dir__.to_s)).strip
   end
 end
